feat: Rename Tests component to Preview in multiple files
