+++ src/views/Brain/Brain.vue
@@ -48,7 +48,7 @@
           :previewActions="previewActions"
         />
         <BrainWarningBar v-if="!routerTunings.brainOn" />
-        <Tests
+        <Preview
           :key="refreshPreviewValue"
           :contentBaseUuid="contentBaseUuid"
           :contentBaseLanguage="contentBase.language"
@@ -77,7 +77,7 @@ import { useStore } from 'vuex';
 import { get } from 'lodash';
 import nexusaiAPI from '../../api/nexusaiAPI';
 import PageContainer from '../../components/PageContainer.vue';
-import Tests from '../repository/content/Tests.vue';
+import Preview from '../repository/content/Preview.vue';
 import RouterMonitoring from './RouterMonitoring/index.vue';
 import RouterAgentsTeam from './RouterAgentsTeam/index.vue';
 import RouterActions from './RouterActions.vue';
@@ -100,7 +100,7 @@ import { useTuningsStore } from '@/store/Tunings';
 export default {
   name: 'Brain',
   components: {
-    Tests,
+    Preview,
     PageContainer,
     RouterMonitoring,
     RouterActions,
